+++ Tests/CrispyWaffle.Tests/Serialization/TestObjects_/SampleTomlClass.cs
@@ -1,0 +1,13 @@
+﻿using System.Diagnostics.CodeAnalysis;
+using CrispyWaffle.Serialization;
+
+namespace CrispyWaffle.Tests.Serialization;
+
+[Serializer(SerializerFormat.Toml)]
+[ExcludeFromCodeCoverage]
+public class SampleTomlClass
+{
+    public string Name { get; set; }
+
+    public int Age { get; set; }
+}