+++ docs/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Add TOML serializer support using Tomlyn - [issue #431](https://github.com/guibranco/CrispyWaffle/issues/431) by [@rodri-oliveira-dev](https://github.com/rodri-oliveira-dev)
+- Add YAML serializer support using YamlDotNet - [issue #432](https://github.com/guibranco/CrispyWaffle/issues/432) by [@rodri-oliveira-dev](https://github.com/rodri-oliveira-dev)
+
 ## Version 10.0 [2025-02-22]
 
 - Drop support to .NET 6 - [pull request #667](https://github.com/guibranco/CrispyWaffle/pull/667)